+++ games_repo/games/bubbles/src/main.rs
@@ -10,7 +10,8 @@ use crate::gfx::SpriteManager;
 use crate::input::InputManager;
 use crate::game::{Game, VIRTUAL_WIDTH, HUD_HEIGHT, PLAY_HEIGHT};
 
-const VERSION: &str = "26.04.11.235";
+#[allow(dead_code)]
+const VERSION: &str = "26.04.11.236";
 
 #[derive(Clone, PartialEq, Debug)]
 enum AppState {